pyproject.toml
[tool.poetry]
name = "thrivewell"
version = "0.1.0"
description = "Companion app using AI"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
pytest = "^8.3.4"
pre-commit = "^4.1.0"
pytest-sugar = "^1.0.0"
pytest-cov = "^6.0.0"
gunicorn = "^23.0.0"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.38"
flask-migrate = "^4.1.0"
poetry-core = "^2.1.1"
dotenv = "^0.9.9"
requests = "^2.32.3"
authlib = "^1.5.0"
flask-testing = "^0.8.1"
flask-mail = "^0.10.0"
flask-wtf = "^1.2.2"
flask-login = "^0.6.3"
itsdangerous = "^2.2.0"
werkzeug = "^3.1.3"
flask-bootstrap = "^3.3.7.1"
email-validator = "^2.2.0"
bs4 = "^0.0.2"
pandas = "^2.2.3"
selenium = "^4.29.0"
webdriver-manager = "^4.0.2"

pgvector = "^0.3.6"
sentence-transformers = "^3.4.1"
openai = "^1.65.2"
langchain-openai = "^0.3.7"
google = "^3.0.0"
google-genai = "^1.4.0"
anthropic = "^0.49.0"
langgraph = "^0.3.5"
redis = "^5.2.1"
gevent = "^24.11.1"
gevent-websocket = "^0.10.1"
aiohttp = "^3.11.13"
eventlet = "^0.39.1"
flask-session = "^0.8.0"
nltk = "^3.9.1"
langchain-community = "^0.3.19"
gpt4all = "^2.8.2"
transformers = "^4.49.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
exclude = ["venv/,  ./application/db/migrations/versions/"]
max-line-length = 79  # Example of another setting

[tool.black]
line-length = 79
